Consolidated AESmodes file
